check that store.fs exists in hasattr.

+++ zarr/storage.py
@@ -119,7 +119,7 @@ def rmdir(store, path: Path = None):
     if hasattr(store, 'rmdir'):
         # pass through
         store.rmdir(path)
-    elif hasattr(store.fs, 'rmdir'):
+    elif hasattr(store, 'fs') and hasattr(store.fs, 'rmdir'):
         store.fs.rmdir(path)
     else:
         # slow version, delete one key at a time
